+++ src/main/control/PhongBanControl.java
@@ -1,6 +1,7 @@
 package src.main.control;
 
 import src.main.boundary.cosovatchat.PhongBanListAdapter;
+import src.main.boundary.cosovatchat.PhongBanView;
 import src.main.entity.CoSoVatChat;
 import src.main.entity.PhongBan;
 
@@ -8,7 +9,7 @@ import java.sql.*;
 import java.util.ArrayList;
 import static src.main.control.ConnectionConfig.connect_to_sql_server;
 public class PhongBanControl {
-
+     public PhongBanView phongBanView;
     public ArrayList<PhongBan> searchPhongBan(String dieuKien) {
         ArrayList<PhongBan> list = new ArrayList<>();
         try{
@@ -60,6 +61,7 @@ public class PhongBanControl {
 
             String sql = String.format("insert into phongban values (N'%s')", item.maPhongBan);
             statement.execute(sql);
+            phongBanView.refreshUI();
             connection.close();
         }
         catch (Exception e){
@@ -72,13 +74,34 @@ public class PhongBanControl {
             Connection connection = connect_to_sql_server();
             Statement statement = connection.createStatement();
 
-            String sql = String.format("delete from hd_phongban where ma_phong_ban = N'%s';", item.maPhongBan);
-            System.out.println(sql);
-            statement.execute(sql);
+            String checkBeforeDelete = String.format("select hoat_dong.ma_hoat_dong\n" +
+                    "from hoat_dong\n" +
+                    "join hd_phongban on hoat_dong.ma_hoat_dong = hd_phongban.ma_hoat_dong\n" +
+                    "where DATEDIFF(dd, ngay_kt, GETDATE()) <= 0 and hd_phongban.ma_phong_ban = N'%s'", item.maPhongBan);
 
-            sql = String.format("delete from phongban where ma_phong_ban = N'%s';", item.maPhongBan);
-            System.out.println(sql);
-            statement.execute(sql);
+            System.out.println(checkBeforeDelete);
+            ResultSet resultSet = statement.executeQuery(checkBeforeDelete);
+
+            String announcement = "Không thể xóa phòng ban này có những hoạt động sau sử dụng:<br/>";
+
+            while(resultSet.next()){
+                announcement = announcement + resultSet.getString(1) + ".<br/>";
+            }
+
+            if(!announcement.equals("Không thể xóa phòng ban này có những hoạt động sau sử dụng:<br/>")){
+                System.out.println(announcement);
+            }
+            else{
+                String sql = String.format("delete from hd_phongban where ma_phong_ban = N'%s';", item.maPhongBan);
+                System.out.println(sql);
+                statement.execute(sql);
+
+                sql = String.format("delete from phongban where ma_phong_ban = N'%s';", item.maPhongBan);
+                System.out.println(sql);
+                phongBanView.refreshUI();
+                statement.execute(sql);
+            }
+
 
             connection.close();
         }
@@ -116,10 +139,14 @@ public class PhongBanControl {
                 System.out.println(sql);
                 statement.execute(sql);
             }
+            phongBanView.refreshUI();
             connection.close();
         }
         catch (Exception e){
             System.out.println(e.getMessage());
         }
+    }
+    public void setPhongBanView(PhongBanView phongBanView) {
+        this.phongBanView = phongBanView;
     }
 }
